# deduplicate package load between AutoBind and Binder