--- a/codegen/config/config.go
+++ b/codegen/config/config.go
@@ -395,22 +395,23 @@ func (c *Config) normalize() error {
 
 	return nil
 }
-
-func (c *Config) Autobind(s *ast.Schema) error {
+func (c *Config) isAutobind(pkg *packages.Package) bool {
+	for _, ab := range c.AutoBind {
+		if strings.HasSuffix(ab, "/...") {
+			abPrefix := strings.TrimSuffix(ab, "/...")
+			if strings.HasPrefix(pkg.PkgPath, abPrefix) {
+				return true
+			}
+		}
+		if pkg.PkgPath == ab {
+			return true
+		}
+	}
+	return false
+}
+func (c *Config) Autobind(s *ast.Schema, ps []*packages.Package) error {
 	if len(c.AutoBind) == 0 {
 		return nil
-	}
-
-	ps, err := packages.Load(&packages.Config{
-		Mode: packages.NeedName |
-			packages.NeedFiles |
-			packages.NeedCompiledGoFiles |
-			packages.NeedImports |
-			packages.NeedTypes |
-			packages.NeedTypesSizes,
-	}, c.AutoBind...)
-	if err != nil {
-		return err
 	}
 
 	for _, t := range s.Types {
@@ -419,6 +420,9 @@ func (c *Config) Autobind(s *ast.Schema) error {
 		}
 
 		for _, p := range ps {
+			if !c.isAutobind(p) {
+				continue
+			}
 			if t := p.Types.Scope().Lookup(t.Name); t != nil {
 				c.Models.Add(t.Name(), t.Pkg().Path()+"."+t.Name())
 				break
@@ -436,6 +440,9 @@ func (c *Config) Autobind(s *ast.Schema) error {
 			}
 
 			for _, p := range ps {
+				if !c.isAutobind(p) {
+					continue
+				}
 				if p.Name != pkg {
 					continue
 				}