--- a/codegen/config/config_test.go
+++ b/codegen/config/config_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/vektah/gqlparser"
 	"github.com/vektah/gqlparser/ast"
+	"golang.org/x/tools/go/packages"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -132,7 +133,19 @@ func TestAutobinding(t *testing.T) {
 		type Message { id: ID }
 	`})
 
-	require.NoError(t, cfg.Autobind(s))
+	ps, err := packages.Load(&packages.Config{
+		Mode: packages.NeedName |
+			packages.NeedFiles |
+			packages.NeedCompiledGoFiles |
+			packages.NeedImports |
+			packages.NeedTypes |
+			packages.NeedTypesSizes |
+			packages.NeedSyntax |
+			packages.NeedTypesInfo,
+	}, cfg.AutoBind...)
+	require.NoError(t, err)
+
+	require.NoError(t, cfg.Autobind(s, ps))
 
 	require.Equal(t, "github.com/99designs/gqlgen/example/scalars/model.Banned", cfg.Models["Banned"].Model[0])
 	require.Equal(t, "github.com/99designs/gqlgen/example/chat.Message", cfg.Models["Message"].Model[0])