This is one of the optimizations suggested in #918.

Improvements against master:

Before:
```
auto bind load 2.684139918s
binder load 19.267647024s
auto bind load 3.140190561s
binder load 19.725400875s
build time 23.085849454s
```
After:
```
combined load 17.965141019s
combined load 17.479315829s
build time 17.721352532s
```

"build time" is the time spent in the build phase of api/generate.go. The first set of loads is part of the models plugin and the second is part of the core functionality.

Note that this reorders InjectBuiltins and AutoBind. I'm not sure what the consequences of that are.

I'm surprised that tests didn't break without https://github.com/99designs/gqlgen/pull/945/commits/8c110b7af3e4685e7678a65884ec18598b2e6c42 but generation in our private repo did. I don't fully understand this code yet. I'd appreciate tips for adding a test for it.

I have:
 - [ ]  Added tests covering the bug / feature (see [testing](https://github.com/99designs/gqlgen/blob/master/TESTING.md))
 - n/a Updated any relevant documentation (see [docs](https://github.com/99designs/gqlgen/tree/master/docs/content))