@vikstrous 2020-01-09
Rebased.
New unscientific benchmark compared to master:

before:
```
46.94user 27.35system 0:31.48elapsed 236%CPU (0avgtext+0avgdata 193760maxresident)k
0inputs+0outputs (1major+401610minor)pagefaults 0swaps
```
after:
```
42.29user 24.44system 0:29.39elapsed 227%CPU (0avgtext+0avgdata 199044maxresident)k
0inputs+0outputs (1major+382901minor)pagefaults 0swaps
```

I also tested on top of #983 and saw the same 3 second improvement. That seems to be how long the package loading is taking for me.