@vikstrous 2020-01-19
New comparison vs master:

**This PR:**
23.82user 9.75system 0:16.15elapsed 207%CPU (0avgtext+0avgdata 189272maxresident)k
0inputs+0outputs (3major+229930minor)pagefaults 0swaps
23.14user 8.50system 0:14.69elapsed 215%CPU (0avgtext+0avgdata 191284maxresident)k
0inputs+0outputs (1major+200711minor)pagefaults 0swaps

**master:**
32.79user 13.08system 0:20.07elapsed 228%CPU (0avgtext+0avgdata 191624maxresident)k
0inputs+0outputs (1major+289754minor)pagefaults 0swaps
33.82user 13.26system 0:20.65elapsed 227%CPU (0avgtext+0avgdata 194580maxresident)k
0inputs+0outputs (1major+257803minor)pagefaults 0swaps