@vektah 2020-02-04
Most of the wiring in this PR has been superseded by https://github.com/99designs/gqlgen/pull/988. 

Getting the same performance gains should be relatively straight forward if we can track down why ReferencedPackages isn't covering all of your deps. Happy to jump on a call (gophers?) if you want a hand to figure this out.